axisym_iter_fit
